+++ sdk/storage/storage-blob/CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugs Fixed
 
+- Changed ContainerClient.createIfNotExists, ContainerClient.deleteIfExists, BlobClient.exists, BlobClient.deleteIfExists, AppendBlobClient.createIfNotExists and PageBlobClient.createIfNotExists to check both header 'x-ms-error-code' and error code in response body. Fixed issue [27913](https://github.com/Azure/azure-sdk-for-js/issues/27913).
+
 ### Other Changes
 
 ## 12.17.0 (2023-11-09)